[AIP-21] Move operators.google_api_to_s3_transfer providers.amazon.aws.operators.google_api_to_s3_transfer
