+++ tests/providers/amazon/aws/operators/test_google_api_to_s3_transfer.py
@@ -23,7 +23,7 @@ from unittest.mock import Mock, patch
 from airflow import models
 from airflow.configuration import load_test_config
 from airflow.models.xcom import MAX_XCOM_SIZE
-from airflow.operators.google_api_to_s3_transfer import GoogleApiToS3Transfer
+from airflow.providers.amazon.aws.operators.google_api_to_s3_transfer import GoogleApiToS3Transfer
 from airflow.utils import db
 
 
@@ -66,9 +66,9 @@ class TestGoogleApiToS3Transfer(unittest.TestCase):
             'dag': None
         }
 
-    @patch('airflow.operators.google_api_to_s3_transfer.GoogleDiscoveryApiHook.query')
-    @patch('airflow.operators.google_api_to_s3_transfer.S3Hook.load_string')
-    @patch('airflow.operators.google_api_to_s3_transfer.json.dumps')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.GoogleDiscoveryApiHook.query')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.S3Hook.load_string')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.json.dumps')
     def test_execute(self, mock_json_dumps, mock_s3_hook_load_string, mock_google_api_hook_query):
         context = {'task_instance': Mock()}
 
@@ -89,9 +89,9 @@ class TestGoogleApiToS3Transfer(unittest.TestCase):
         context['task_instance'].xcom_pull.assert_not_called()
         context['task_instance'].xcom_push.assert_not_called()
 
-    @patch('airflow.operators.google_api_to_s3_transfer.GoogleDiscoveryApiHook.query')
-    @patch('airflow.operators.google_api_to_s3_transfer.S3Hook.load_string')
-    @patch('airflow.operators.google_api_to_s3_transfer.json.dumps')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.GoogleDiscoveryApiHook.query')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.S3Hook.load_string')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.json.dumps')
     def test_execute_with_xcom(self, mock_json_dumps, mock_s3_hook_load_string, mock_google_api_hook_query):
         context = {'task_instance': Mock()}
         xcom_kwargs = {
@@ -124,10 +124,13 @@ class TestGoogleApiToS3Transfer(unittest.TestCase):
             value=mock_google_api_hook_query.return_value
         )
 
-    @patch('airflow.operators.google_api_to_s3_transfer.GoogleDiscoveryApiHook.query')
-    @patch('airflow.operators.google_api_to_s3_transfer.S3Hook.load_string')
-    @patch('airflow.operators.google_api_to_s3_transfer.json.dumps')
-    @patch('airflow.operators.google_api_to_s3_transfer.sys.getsizeof', return_value=MAX_XCOM_SIZE)
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.GoogleDiscoveryApiHook.query')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.S3Hook.load_string')
+    @patch('airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.json.dumps')
+    @patch(
+        'airflow.providers.amazon.aws.operators.google_api_to_s3_transfer.sys.getsizeof',
+        return_value=MAX_XCOM_SIZE
+    )
     def test_execute_with_xcom_exceeded_max_xcom_size(
         self,
         mock_sys_getsizeof,
